Updated and fixed delete account logic
